+++ src/app/user/user-details/user-details.component.ts
@@ -11,11 +11,12 @@ import { MatTooltipModule } from '@angular/material/tooltip';
 import { ReactiveFormsModule, FormControl, FormGroup } from '@angular/forms';
 import {MatSelectModule} from '@angular/material/select';
 import { TranslateModule } from '@ngx-translate/core';
+import { UpperCasePipe } from '@angular/common';
 
 @Component({
   selector: 'user-details',
   standalone: true,
-  imports: [AsyncPipe, MatSelectModule, MatIconModule, TranslateModule, MatButtonModule, MatTooltipModule, ReactiveFormsModule],
+  imports: [AsyncPipe, MatSelectModule, MatIconModule, UpperCasePipe, TranslateModule, MatButtonModule, MatTooltipModule, ReactiveFormsModule],
   templateUrl: './user-details.component.html',
   styleUrl: './user-details.component.scss'
 })
