+++ src/app/user/user-list-item/user-list-item.component.html
@@ -7,7 +7,7 @@
     </div>
     <div class="role-container">
       @for (role of user.roles; track $index) {
-        <p class="role">{{ role | translate }}</p>
+        <p class="role">{{ "ROLES." + role | uppercase | translate }}</p>
       }
     </div>
   }
